refactor: restructure the organization package
