+++ internal/plugin/service/organization/project/organization_group_project.go
@@ -1,4 +1,4 @@
-package organization
+package project
 
 import (
 	"context"
