+++ internal/plugin/service/organization/usergroup/organization_user_group_member.go
@@ -1,4 +1,4 @@
-package organization
+package usergroup
 
 import (
 	"context"
